server/requirements.txt
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.4.1
boto==2.40.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
distlib==0.3.3
filelock==3.4.0
Flask==0.10.1
Flask-Cors==3.0.2
futures==3.0.5
google-api-core==2.3.0
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
itsdangerous==0.24
Jinja2==2.8
lxml==4.2.5
MarkupSafe==0.23
oauth2client==2.2.0
Paste==3.5.0
platformdirs==2.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
redis==2.10.5
requests==2.26.0
rsa==3.4.2
six==1.10.0
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
webapp2==2.5.2
WebOb==1.8.7
Werkzeug==0.11.9
